create plan and view plan list

+++ src/main/java/org/nvtai/ProductionSchedulingSystem/entity/Role.java
@@ -14,10 +14,9 @@ import java.util.Set;
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class Role {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long rid;
+    Integer rid;
 
     String rname;
